Added package checks, re-enabled sw-install.
sw-install will now ensure that pacaur is installed as it can be run without root, and will also check both that every package in aur_packages.txt exist, but are not already installed before attempting to install. This will allow the script to be run manually for a user to install anything on the list that is not already installed, but does not need to *reinstall* everything that is on the list and already installed. Additionally, changed the actual calls sw-install makes from the testing/placeholder "echo" calls to the proper calls.